# web/main-ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "inventoryzing",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/objects/{object_id}/properties": {
      "get": {
        "summary": "Object Properties",
        "operationId": "object_properties_api_objects__object_id__properties_get",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Object Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertyValue"
                  },
                  "title": "Response Object Properties Api Objects  Object Id  Properties Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/types/{type_id}/properties": {
      "get": {
        "summary": "Type Properties",
        "operationId": "type_properties_api_types__type_id__properties_get",
        "parameters": [
          {
            "name": "type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Type Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertyValue"
                  },
                  "title": "Response Type Properties Api Types  Type Id  Properties Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JsonValue": {},
...
      "PropertyValue": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "date",
              "datetime",
              "number",
              "integer",
              "decimal",
              "boolean",
              "quantity"
            ],
            "title": "Type",
            "default": "text"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "example": {
            "type": "string",
            "title": "Example",
            "default": ""
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "quantity_dimension": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity Dimension"
          },
          "canonical_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Unit"
          },
          "allowed_units": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Units"
          },
          "schema_revision": {
            "type": "integer",
            "title": "Schema Revision",
            "default": 1
          },
          "editable": {
            "type": "boolean",
            "title": "Editable",
            "default": false
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          },
          "declared_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Declared On"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "$ref": "#/components/schemas/JsonValue"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "formatted_value": {
            "type": "string",
            "title": "Formatted Value",
            "default": ""
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "missing",
              "unavailable"
            ],
            "title": "Status",
            "default": "missing"
          },
          "effective_state": {
            "type": "string",
            "enum": [
              "value",
              "unset",
              "missing"
            ],
            "title": "Effective State",
            "default": "missing"
          },
          "local_state": {
            "type": "string",
            "enum": [
              "inherit",
              "value",
              "unset"
            ],
            "title": "Local State",
            "default": "inherit"
          },
          "source_kind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "object",
                  "object_type"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Kind"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "applicability_source_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicability Source Id"
          },
          "applicability_source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applicability Source Name"
          },
          "declared_directly": {
            "type": "boolean",
            "title": "Declared Directly",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "provider"
        ],
        "title": "PropertyValue"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}